FAI-14236 Move azurepipeline models to faros-airbyte-common
